refactor: move internal/cloud to pkg/cloud
Makes the cloud provider package importable from external modules
(e.g. onctl-platform SaaS backend). The internal/ restriction only
blocked the Go compiler, not source visibility — the repo is public
either way, so there is no security benefit to keeping it internal.

All import paths updated; build and tests confirmed clean.
