+++ pkg/cloud/cloud_test.go
